+++ src/app/(protected)/settings/_components/SectionDivider.tsx
@@ -1,0 +1,3 @@
+export default function SectionDivider() {
+    return <div className="h-4 w-full shrink-0 bg-[#020202]" />
+}
